refactor: create components dir, move svg to images
